chore: address issue #148
