+++ packages/seisai-engine/src/seisai_engine/pipelines/fbpick/physics/physical_center.py
@@ -18,9 +18,6 @@ from .feasible import FeasibleBandResult, compute_velocity_t0_band_from_arrays
 from .geometry import (
     CoarseGeometry,
     SourceGroup,
-    build_source_groups,
-    is_source_xy_degenerate,
-    load_coarse_geometry_from_npz,
     select_nearest_source_groups,
     signed_offset_side_from_geometry,
     split_offset_gap_segments,
@@ -36,6 +33,15 @@ from .physical_center_fallback import (
     _emit_fallback_all_and_done,
     _finalize_result_with_pending_trend_fallback,
     _PendingTrendFallback,
+)
+from .physical_center_geometry import (
+    _signed_offset_side_labels,
+    _SignedOffsetSideLabels,
+    build_physical_center_geometry_context,
+    build_physical_center_offset_context,
+    build_physical_center_source_group_build,
+    build_physical_center_source_group_context,
+    load_physical_center_geometry,
 )
 from .physical_center_types import (
     PHYSICAL_MODEL_FAILURE_FIT_FAILED,
@@ -137,12 +143,6 @@ class _ObservationPlan:
     segment_id: int
     side: int
     relaxed: bool
-
-
-@dataclass(frozen=True)
-class _SignedOffsetSideLabels:
-    side: np.ndarray
-    finite: np.ndarray
 
 
 @dataclass(frozen=True)
@@ -423,40 +423,6 @@ def _side_slot(side: int) -> int:
         msg = 'side must be -1, 0, or 1'
         raise ValueError(msg)
     return value + 1
-
-
-def _signed_offset_side_labels(
-    signed_offset_m: np.ndarray,
-    *,
-    zero_tol_m: float = 1.0e-6,
-    finite_mask: np.ndarray | None = None,
-) -> _SignedOffsetSideLabels:
-    signed = np.asarray(signed_offset_m, dtype=np.float32)
-    if signed.ndim != 1:
-        msg = 'signed_offset_m must be 1D'
-        raise ValueError(msg)
-
-    zero_tol = float(zero_tol_m)
-    if zero_tol < 0.0 or not np.isfinite(zero_tol):
-        msg = 'zero_tol_m must be finite and >= 0'
-        raise ValueError(msg)
-
-    finite = np.isfinite(signed)
-    if finite_mask is not None:
-        mask = np.asarray(finite_mask, dtype=np.bool_)
-        if mask.shape != signed.shape:
-            msg = 'finite_mask must have the same shape as signed_offset_m'
-            raise ValueError(msg)
-        finite = finite & mask
-
-    side = np.zeros(signed.shape, dtype=np.int8)
-    signed_valid = np.asarray(signed[finite], dtype=np.float64)
-    side[finite] = np.where(
-        np.abs(signed_valid) <= zero_tol,
-        0,
-        np.where(signed_valid < 0.0, -1, 1),
-    ).astype(np.int8)
-    return _SignedOffsetSideLabels(side=side, finite=finite)
 
 
 def _filtered_obs_key(
@@ -3052,114 +3018,19 @@ def _adaptive_refit_triggered(
     return bool(resid_trigger or shift_trigger or insufficient_trigger)
 
 
-def _table_offset_abs_m(table: CoarsePickTable, *, n_traces: int) -> np.ndarray:
-    offset_m = _as_vector(
-        'table.offset_m',
-        table.offset_m,
-        n_traces=n_traces,
-        dtype=np.float32,
-    )
-    return np.abs(offset_m).astype(np.float32, copy=False)
-
-
-def _build_table_source_groups(
-    table: CoarsePickTable,
-    *,
-    n_traces: int,
-) -> tuple[SourceGroup, ...]:
-    shot_id = _as_vector(
-        'table.shot_id',
-        table.shot_id,
-        n_traces=n_traces,
-        dtype=np.int32,
-    )
-    groups: list[SourceGroup] = []
-    seen: set[int] = set()
-    for shot in shot_id.tolist():
-        shot_int = int(shot)
-        if shot_int in seen:
-            continue
-        seen.add(shot_int)
-        trace_indices = np.flatnonzero(shot_id == np.int32(shot_int)).astype(
-            np.int64,
-            copy=False,
-        )
-        group_id = len(groups)
-        groups.append(
-            SourceGroup(
-                group_id=group_id,
-                source_key_x=shot_int,
-                source_key_y=0,
-                source_x_m=float(group_id),
-                source_y_m=0.0,
-                trace_indices=trace_indices,
-            )
-        )
-    return tuple(groups)
-
-
-def _load_source_group_geometry_from_npz(
-    coarse_npz: Mapping[str, np.ndarray],
-    *,
-    n_traces: int,
-) -> CoarseGeometry | None:
-    if 'source_x_m' not in coarse_npz or 'source_y_m' not in coarse_npz:
-        return None
-    try:
-        source_x_m = _as_vector(
-            'source_x_m',
-            coarse_npz['source_x_m'],
-            n_traces=n_traces,
-            dtype=np.float32,
-        )
-        source_y_m = _as_vector(
-            'source_y_m',
-            coarse_npz['source_y_m'],
-            n_traces=n_traces,
-            dtype=np.float32,
-        )
-        if 'geometry_valid_mask' in coarse_npz:
-            geometry_valid_mask = _as_bool_vector(
-                'geometry_valid_mask',
-                coarse_npz['geometry_valid_mask'],
-                n_traces=n_traces,
-            )
-        else:
-            geometry_valid_mask = np.ones((n_traces,), dtype=np.bool_)
-    except (TypeError, ValueError):
-        return None
-
-    geometry_valid_mask = (
-        geometry_valid_mask
-        & np.isfinite(source_x_m)
-        & np.isfinite(source_y_m)
-    ).astype(np.bool_, copy=False)
-    zeros = np.zeros((n_traces,), dtype=np.float32)
-    return CoarseGeometry(
-        source_x_m=source_x_m,
-        source_y_m=source_y_m,
-        receiver_x_m=zeros,
-        receiver_y_m=zeros.copy(),
-        offset_abs_geom_m=zeros.copy(),
-        geometry_valid_mask=geometry_valid_mask,
-        offset_signed_geom_m=None,
-    )
-
-
 def preflight_geometry_two_piece_fallback(
     *,
     coarse_npz: Mapping[str, np.ndarray],
     table: CoarsePickTable,
     cfg: PhysicsLiteConfig,
 ) -> PhysicalCenterFallbackPreflight:
-    n = int(table.n_traces)
-    use_geometry_offset = bool(cfg.physical_trend.use_geometry_offset)
-    try:
-        geometry = load_coarse_geometry_from_npz(coarse_npz, n_traces=n)
-    except (KeyError, TypeError, ValueError):
-        geometry = None
-
-    if use_geometry_offset and geometry is None:
+    geometry_context = build_physical_center_geometry_context(
+        coarse_npz=coarse_npz,
+        table=table,
+        cfg=cfg,
+        include_offsets=False,
+    )
+    if geometry_context.geometry_required_missing:
         return PhysicalCenterFallbackPreflight(
             status='geometry_invalid',
             reason='geometry_invalid',
@@ -3168,53 +3039,28 @@ def preflight_geometry_two_piece_fallback(
             groups=None,
         )
 
-    source_grouping_invalid = False
-    groups: tuple[SourceGroup, ...] = ()
-    source_group_geometry = geometry
-    if source_group_geometry is None and not use_geometry_offset:
-        source_group_geometry = _load_source_group_geometry_from_npz(
-            coarse_npz,
-            n_traces=n,
-        )
-    if source_group_geometry is not None:
-        coord_group_tol_m = float(cfg.physical_trend.coord_group_tol_m)
-        source_xy_degenerate = is_source_xy_degenerate(
-            source_group_geometry,
-            table=table,
-            coord_group_tol_m=coord_group_tol_m,
-        )
-        if source_xy_degenerate and use_geometry_offset:
-            source_grouping_invalid = True
-        if not source_xy_degenerate:
-            groups = build_source_groups(
-                source_group_geometry,
-                coord_group_tol_m=coord_group_tol_m,
-            )
-    if len(groups) == 0 and not use_geometry_offset:
-        groups = _build_table_source_groups(table, n_traces=n)
-
-    if source_grouping_invalid:
+    if geometry_context.source_grouping_invalid:
         return PhysicalCenterFallbackPreflight(
             status='geometry_invalid',
             reason='source_xy_degenerate',
             fallback_mode=str(cfg.physical_runtime.geometry_invalid_fallback),
-            geometry_loaded=geometry is not None,
-            groups=len(groups),
+            geometry_loaded=geometry_context.geometry is not None,
+            groups=len(geometry_context.groups),
         )
-    if len(groups) == 0:
+    if len(geometry_context.groups) == 0:
         return PhysicalCenterFallbackPreflight(
             status='geometry_invalid',
             reason='source_group_empty',
             fallback_mode=str(cfg.physical_runtime.group_invalid_fallback),
-            geometry_loaded=geometry is not None,
+            geometry_loaded=geometry_context.geometry is not None,
             groups=0,
         )
     return PhysicalCenterFallbackPreflight(
         status=None,
         reason=None,
         fallback_mode=None,
-        geometry_loaded=geometry is not None,
-        groups=len(groups),
+        geometry_loaded=geometry_context.geometry is not None,
+        groups=len(geometry_context.groups),
     )
 
 
@@ -3321,10 +3167,7 @@ def build_geometry_two_piece_physical_center(
         if runtime_diagnostics is not None
         else nullcontext()
     ):
-        try:
-            geometry = load_coarse_geometry_from_npz(coarse_npz, n_traces=n)
-        except (KeyError, TypeError, ValueError):
-            geometry = None
+        geometry = load_physical_center_geometry(coarse_npz, n_traces=n)
     reporter.emit(
         'physical-center.stage_done',
         **context,
@@ -3350,42 +3193,15 @@ def build_geometry_two_piece_physical_center(
             runtime_diagnostics=runtime_diagnostics,
         )
 
-    segment_by_offset_sign = bool(cfg.physical_trend.segment_by_offset_sign)
-    offset_signed_labels = None
-    if use_geometry_offset:
-        offset_abs_m = _as_vector(
-            'geometry.offset_abs_geom_m',
-            geometry.offset_abs_geom_m,
-            n_traces=n,
-            dtype=np.float32,
-        )
-        if segment_by_offset_sign and geometry.offset_signed_geom_m is not None:
-            offset_signed_m = _as_vector(
-                'geometry.offset_signed_geom_m',
-                geometry.offset_signed_geom_m,
-                n_traces=n,
-                dtype=np.float32,
-            )
-            offset_signed_labels = _signed_offset_side_labels(
-                offset_signed_m,
-                finite_mask=geometry.geometry_valid_mask,
-            )
-        else:
-            offset_signed_m = None
-        offset_source = PHYSICAL_OFFSET_SOURCE_GEOMETRY
-    else:
-        offset_abs_m = _table_offset_abs_m(table, n_traces=n)
-        if segment_by_offset_sign:
-            offset_signed_m = _as_vector(
-                'table.offset_m',
-                table.offset_m,
-                n_traces=n,
-                dtype=np.float32,
-            )
-            offset_signed_labels = _signed_offset_side_labels(offset_signed_m)
-        else:
-            offset_signed_m = None
-        offset_source = PHYSICAL_OFFSET_SOURCE_HEADER
+    offset_context = build_physical_center_offset_context(
+        geometry=geometry,
+        table=table,
+        cfg=cfg,
+    )
+    offset_abs_m = offset_context.offset_abs_m
+    offset_signed_m = offset_context.offset_signed_m
+    offset_signed_labels = offset_context.offset_signed_labels
+    offset_source = offset_context.offset_source
 
     reporter.emit('physical-center.stage_start', **context, stage='source_grouping')
     stage_start = time.perf_counter()
@@ -3394,32 +3210,13 @@ def build_geometry_two_piece_physical_center(
         if runtime_diagnostics is not None
         else nullcontext()
     ):
-        source_grouping_invalid = False
-        source_groups_from_geometry = False
-        groups: tuple[SourceGroup, ...] = ()
-        source_group_geometry = geometry
-        if source_group_geometry is None and not use_geometry_offset:
-            source_group_geometry = _load_source_group_geometry_from_npz(
-                coarse_npz,
-                n_traces=n,
-            )
-        if source_group_geometry is not None:
-            coord_group_tol_m = float(cfg.physical_trend.coord_group_tol_m)
-            source_xy_degenerate = is_source_xy_degenerate(
-                source_group_geometry,
-                table=table,
-                coord_group_tol_m=coord_group_tol_m,
-            )
-            if source_xy_degenerate and use_geometry_offset:
-                source_grouping_invalid = True
-            if not source_xy_degenerate:
-                groups = build_source_groups(
-                    source_group_geometry,
-                    coord_group_tol_m=coord_group_tol_m,
-                )
-                source_groups_from_geometry = len(groups) > 0
-        if len(groups) == 0 and not use_geometry_offset:
-            groups = _build_table_source_groups(table, n_traces=n)
+        source_group_build = build_physical_center_source_group_build(
+            coarse_npz=coarse_npz,
+            geometry=geometry,
+            table=table,
+            cfg=cfg,
+        )
+        groups = source_group_build.groups
     reporter.emit(
         'physical-center.stage_done',
         **context,
@@ -3428,7 +3225,7 @@ def build_geometry_two_piece_physical_center(
         groups=len(groups),
     )
 
-    if source_grouping_invalid:
+    if source_group_build.source_grouping_invalid:
         return _emit_fallback_all_and_done(
             status='geometry_invalid',
             reason='source_xy_degenerate',
@@ -3474,12 +3271,15 @@ def build_geometry_two_piece_physical_center(
         if runtime_diagnostics is not None
         else nullcontext()
     ):
-        groups_by_id = {int(group.group_id): group for group in groups}
-        group_id_by_trace = np.full((n,), -1, dtype=np.int32)
-        for group in groups:
-            group_id_by_trace[
-                np.asarray(group.trace_indices, dtype=np.int64)
-            ] = np.int32(group.group_id)
+        source_group_context = build_physical_center_source_group_context(
+            source_group_build=source_group_build,
+            n_traces=n,
+        )
+        groups_by_id = source_group_context.groups_by_id
+        group_id_by_trace = source_group_context.group_id_by_trace
+        source_groups_from_geometry = (
+            source_group_context.source_groups_from_geometry
+        )
     reporter.emit(
         'physical-center.stage_done',
         **context,
